// src/app/common-objects/place/dto/associated-area.dto.ts
export class AssociatedAreaDTO {
	areaAssociationID: string;
	associationTypeID: string;
	associationTypeName: string;
	subjectAreaID: string;
	subjectAreaName: string;
	associatedAreaID: string;
	associatedAreaName: string;
	areaID: string;
	areaCode: string;
	areaName: string;
	areaDescription;
	areaPermanentRecordIndicator
	areaStatusID: string;
	areaStatusName: string;
}